+++ PgRoutiner/Builder/CodeBuilder/Crud/CrudDeleteReturningCode.cs
@@ -55,7 +55,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.PkParams.Select(p => $"{I5}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I4}.Single();");
+            Class.AppendLine($"{I4}.{settings.SingleLinqMethod}();");
             Class.AppendLine($"{I2}}}");
             AddMethod(name, actualReturns, true);
         }
@@ -77,7 +77,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.PkParams.Select(p => $"{I5}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I4}.SingleAsync();");
+            Class.AppendLine($"{I4}.{settings.SingleLinqMethod}Async();");
             Class.AppendLine($"{I2}}}");
             AddMethod(name, actualReturns, false);
         }
@@ -97,7 +97,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.PkParams.Select(p => $"{I5}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I3}.Single();");
+            Class.AppendLine($"{I3}.{settings.SingleLinqMethod}();");
             AddMethod(name, actualReturns, true);
         }
 
@@ -117,7 +117,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.PkParams.Select(p => $"{I5}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I3}.SingleAsync();");
+            Class.AppendLine($"{I3}.{settings.SingleLinqMethod}Async();");
             AddMethod(name, actualReturns, false);
         }