+++ PgRoutiner/Builder/CodeBuilder/Crud/CrudUpdateReturningCode.cs
@@ -65,7 +65,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I5}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I4}.Single();");
+            Class.AppendLine($"{I4}.{settings.SingleLinqMethod}();");
             Class.AppendLine($"{I2}}}");
             AddMethod(name, actualReturns, true);
         }
@@ -87,7 +87,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I5}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I4}.SingleAsync();");
+            Class.AppendLine($"{I4}.{settings.SingleLinqMethod}Async();");
             Class.AppendLine($"{I2}}}");
             AddMethod(name, actualReturns, false);
         }
@@ -107,7 +107,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I4}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I3}.Single();");
+            Class.AppendLine($"{I3}.{settings.SingleLinqMethod}();");
             AddMethod(name, actualReturns, true);
         }
 
@@ -127,7 +127,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I4}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I3}.SingleAsync();");
+            Class.AppendLine($"{I3}.{settings.SingleLinqMethod}Async();");
             AddMethod(name, actualReturns, false);
         }