+++ PgRoutiner/Settings/BuildFormatedSettings.cs
@@ -136,6 +136,7 @@ namespace PgRoutiner
             AddEntry(nameof(MinNormVersion), Value.MinNormVersion);
             AddEntry(nameof(SourceHeader), Value.SourceHeader);
             AddEntry(nameof(Ident), Value.Ident);
+            AddEntry(nameof(SingleLinqMethod), Value.SingleLinqMethod);
 
             sb.AppendLine();
             AddSectionComment(